colour_math_ng::Chroma: Either -> Neither redefins eq() and cmp()
all hue_tests test broken and need to be fixed. Everywhere Chroma
types are created now need to be specific. Chroma::Neither() means
not a shade or a tint i.e. Chroma::ZERO, Chroma::ONE and the chroma
for colours that are neither darker or lighter than the maximum
chroma RGB for their hue.

Signed-off-by: Peter Williams <pwil3058@gmail.com>